requirements.txt
absl-py==2.0.0
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
clearml==1.13.1
filelock==3.12.4
fire==0.5.0
frozenlist==1.4.0
fsspec==2023.9.2
furl==2.1.3
google-auth==2.23.3
google-auth-oauthlib==1.0.0
grpcio==1.59.0
h5py==3.10.0
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
lightning-utilities==0.9.0
Markdown==3.5
MarkupSafe==2.1.3
monai==1.3.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
orderedmultidict==1.0.1
packaging==23.2
pandas==1.4.4
pathlib2==2.3.7.post1
Pillow==10.1.0
protobuf==4.24.4
psutil==5.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.4.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytorch-lightning==2.1.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.6
rsa==4.9
scikit-learn==1.3.1
scipy==1.11.3
six==1.16.0
sympy==1.12
tensorboard==2.14.1
tensorboard-data-server==0.7.1
termcolor==2.3.0
threadpoolctl==3.2.0
torch==2.1.0
torchmetrics==1.2.0
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
Werkzeug==3.0.0
yarl==1.9.2
zipp==3.17.0
